Update app_front.py
Render was only running app_front.py when the app was deployed. The changes here have the code ping the SQL table each time there is a visitor. A consideration for the future is to revert to before but have GitHub Actions run app_front.py once a week after it runs comiccode.py